requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
bitarray==2.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
configparser==6.0.0
coverage==7.3.4
cytoolz==0.12.2
discord-webhook==1.3.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
linecache2==1.0.0
lru-dict==1.2.0
mock==5.1.0
multidict==6.0.4
oauthlib==3.2.2
outcome==1.3.0.post0
parsimonious==0.9.0
protobuf==4.25.1
pycparser==2.21
pycryptodome==3.19.0
PySocks==1.7.1
python-dotenv==1.0.0
pyunormalize==15.1.0
pywin32==306
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.0.0
rpds-py==0.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toolz==0.12.0
traceback2==1.4.0
trio==0.24.0
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.9.0
unittest2==1.1.0
urllib3==2.1.0
web3==6.13.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
